requirements.txt
black==22.3.0
click==8.1.3
pytest==7.4.0
pytest-cov==4.0.0
pylint==2.15.3
boto3==1.24.87
requests
ruff

# Databricks connection & environment variables handling
databricks-sql-connector
python-dotenv